+++ src/tests/test_orca.cc
@@ -226,6 +226,7 @@ BOOST_AUTO_TEST_CASE(input_generation_version_4_0_1) {
            << "<orca>\n"
            << "<method></method>\n"
            << "<scf>GUESS PMODEL</scf>\n"
+           << "<maxcore>3000</maxcore>\n"
            << "</orca>\n"
            << "</package>";
   defaults.close();
@@ -261,7 +262,7 @@ BOOST_AUTO_TEST_CASE(input_generation_version_4_0_1) {
   BOOST_CHECK_EQUAL(inp.substr(index1, index2 - index1),
                     "%basis\nGTOName =\"system.bas\";\n");
 
-  // check basis section
+  // check scf section multiline
   index1 = inp.find("%scf");
   index2 = inp.find("end", index1);
   BOOST_CHECK_EQUAL(inp.substr(index1, index2 - index1),
@@ -270,6 +271,12 @@ BOOST_AUTO_TEST_CASE(input_generation_version_4_0_1) {
   // Check method
   index1 = inp.find("!");
   BOOST_CHECK_EQUAL(inp.substr(index1), "! DFT pbe0   \n");
+
+  // Check singleline orca kewords
+  index1 = inp.find("%maxcore");
+  index2 = inp.find("\n", index1);
+  std::cout << "\ninp: " << inp.substr(index1) << "\n";
+  BOOST_CHECK_EQUAL(inp.substr(index1, index2 - index1), "%maxcore 3000");
 }
 
 BOOST_AUTO_TEST_SUITE_END()